# CUDA: Use __grid_constant__ kernel argument launch  mode if available