We now can use up to 32kB of size for parameters for global function, and can have the compiler manage the use of constant cache for that argument, through the __grid_constant__ parameter. Only available for CC 7.0 or later and since CUDA 12.2 with NVCC (not clang as CUDA compiler).

I did some testing with the following code:

```c++
template<size_t N>
struct functor {
  Kokkos::View<int***> a,b,c;
  float data[N];

  functor(Kokkos::View<int***> a_, Kokkos::View<int***> b_, Kokkos::View<int***> c_, float* data_):a(a_),b(b_),c(c_) {
    for(int i = 0; i<N; i++) data[i] = data_[i];
  }

  KOKKOS_FUNCTION
  void operator()(int i) const {
    for(int j=0; j<a.extent(1); j++)
      for(int k=0; k<a.extent(2); k++)
        a(i,j,k) += b(i,j,k) + c(i,j,k) + data[k % N];
  }
};

template<size_t K>
void foo(int N, int M, int KK, int R) {
  Kokkos::View<int***> a("A", N, M, KK);
  Kokkos::View<int***> b("A", N, M, KK);
  Kokkos::View<int***> c("A", N, M, KK);

  float data[K];
  for(int k=0; k<K; k++) data[k] = k;
  Kokkos::deep_copy(b,2);
  Kokkos::deep_copy(c,3);

  functor<K> f(a,b,c,data);

  Kokkos::parallel_for(N, f);

  Kokkos::Timer timer;
  for(int r = 0; r<R; r++) Kokkos::parallel_for(N, f);
  Kokkos::fence();
  double time = timer.seconds();

  size_t errors;
  Kokkos::parallel_reduce(N, KOKKOS_LAMBDA(int i, size_t& lsum) {
    for(int j=0; j<a.extent(1); j++)
      for(int k=0; k<a.extent(2); k++)
        if(a(i,j,k) != (R+1)*(2 + 3 + k%K)) lsum++;
  }, errors);
  printf("N=%i M=%i K=%lu KK=%i R=%i Time: %lf us Errors: %lu\n",N,M,K,KK,R,time * 1.e6/ R, errors);
}
```

New:
```
N=1 M=1 K=1 KK=1 R=20 Time: 3.780050 us Errors: 0
N=1 M=1 K=4000 KK=1 R=20 Time: 9.266150 us Errors: 0
N=1 M=1 K=40000 KK=1 R=20 Time: 41.734450 us Errors: 0

N=1 M=2 K=1 KK=1000 R=20 Time: 180.949000 us Errors: 0
N=1 M=2 K=4000 KK=1000 R=20 Time: 186.581900 us Errors: 0
N=1 M=2 K=40000 KK=1000 R=20 Time: 478.332450 us Errors: 0

N=200000 M=1 K=1 KK=1 R=20 Time: 5.617950 us Errors: 0
N=200000 M=1 K=4000 KK=1 R=20 Time: 9.809200 us Errors: 0
N=200000 M=1 K=40000 KK=1 R=20 Time: 41.711850 us Errors: 0

N=200000 M=2 K=1 KK=100 R=20 Time: 876.692750 us Errors: 0
N=200000 M=2 K=4000 KK=100 R=20 Time: 913.998600 us Errors: 0
N=200000 M=2 K=40000 KK=100 R=20 Time: 1083.674100 us Errors: 0
```

Old:
```
N=1 M=1 K=1 KK=1 R=20 Time: 3.788900 us Errors: 0
N=1 M=1 K=4000 KK=1 R=20 Time: 15.624350 us Errors: 0
N=1 M=1 K=40000 KK=1 R=20 Time: 42.278350 us Errors: 0

N=1 M=2 K=1 KK=1000 R=20 Time: 180.940250 us Errors: 0
N=1 M=2 K=4000 KK=1000 R=20 Time: 199.479100 us Errors: 0
N=1 M=2 K=40000 KK=1000 R=20 Time: 478.678350 us Errors: 0

N=200000 M=1 K=1 KK=1 R=20 Time: 5.784100 us Errors: 0
N=200000 M=1 K=4000 KK=1 R=20 Time: 18.831600 us Errors: 0
N=200000 M=1 K=40000 KK=1 R=20 Time: 42.156050 us Errors: 0

N=200000 M=2 K=1 KK=100 R=20 Time: 874.896100 us Errors: 0
N=200000 M=2 K=4000 KK=100 R=20 Time: 924.781400 us Errors: 0
N=200000 M=2 K=40000 KK=100 R=20 Time: 1083.891950 us Errors: 0
```

As you can see the difference is there for the mid-size functors. The smallest doesn't see a difference, and the biggest doesn't (the K parameter is essentially the size of a stored array in the functor). 
But for the midsize functor (16kB) the minimum latency goes from 15us to 9us and the runtime for 200k iterations goes from 18.8us to 9.8us. So quite a bit of improvement.