@masterleinad 2025-10-07
> Only available for CC 7.0 or later and since CUDA 12.2 with NVCC (not clang as CUDA compiler).

Why wouldn't it work with clang? https://clang.llvm.org/docs/AttributeReference.html#grid-constant shows support.