@crtrott 2025-10-08
@romintomasetti yes the reason you don't see much performance difference is that we only use local launch for small functors where it doesn't matter (<200 bytes) right now. If you get larger it does.

Here is the same test from above for some selected output (the functor size is K*4 + 3*sizeof(View<int***>)
N=200000 M=2 K=10 KK=10 R=20 Time: 99.164650 us Errors: 0
N=200000 M=2 K=100 KK=100 R=20 Time: 905.883750 us Errors: 0
N=200000 M=2 K=1000 KK=1000 R=20 Time: 10641.231300 us Errors: 0

Here is the same with grid constant removed:
N=200000 M=2 K=10 KK=10 R=20 Time: 99.016450 us Errors: 0
N=200000 M=2 K=100 KK=100 R=20 Time: 1154.456150 us Errors: 0
N=200000 M=2 K=1000 KK=1000 R=20 Time: 13471.315150 us Errors: 0

As you can see for the functor which is about 500 bytes and the one which is 4000, the argument makes about 25% speed difference even though nominally this kernel should be completely bandwidth limited.
For the functor where the size is <200 bytes it doesn't make a difference.